requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dash==3.2.0
Flask==3.1.2
Flask-SocketIO==5.5.1
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
narwhals==2.6.0
nest-asyncio==1.6.0
packaging==25.0
plotly==6.3.1
pymodbus==3.11.3
pyserial==3.5
python-engineio==4.12.3
python-socketio==5.14.1
PyYAML==6.0.3
requests==2.32.5
retrying==1.4.2
simple-websocket==1.1.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
